segurança desabilitada e CRUD pessoa feito, faltando apenas lidar com as exceções

+++ src/main/java/br/ufrn/ePET/repository/PessoaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.ufrn.ePET.models.Pessoa;
 
-public interface PessoaRepository extends JpaRepository<Pessoa, Integer>{
+public interface PessoaRepository extends JpaRepository<Pessoa, Long>{
 	Pessoa findByNome(String nome);
 }
